refactor: Extract `ams` urls
